pyproject.toml
[tool.poetry]
name = "gpt3-api"
version = "0.1.0"
description = "pack_name descr "
authors = ["ffreemt"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/ffreemt/gpt3-api"
include = ["LICENSE"]

[tool.poetry.dependencies]
python = ">=3.7.1,<4.0"
logzero = "^1.7.0"
PyInquirer = "^1.0.3"
pydantic = "^1.8.2"
openai = "^0.11.0"
python-dotenv = "^0.19.2"
environs = "^9.3.5"

[tool.poe.executor]
type = "poetry"

[tool.poe.tasks]
memo = "echo poe test"
test = "pytest tests"
build = "poetry build"
_publish = "poetry publish"
release = ["test", "build", "_publish"]
style = { cmd = "flake8 gpt3_api" }
docstyle = { cmd = "pydocstyle gpt3_api" }
format = "black tests gpt3_api"
lint = { cmd = "pylint gpt3_api" }
tunnel = {cmd ="ssh -CN ip_or_hostname_defined_in_hosts -L 9091:127.0.0.1:9091"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.9.2"
pep257 = "^0.7.0"
tbump = "^6.3.2"
poethepoet = "^0.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
